requirements.txt
flask>=2.3.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
groq>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
dateparser>=1.2.0
pypdf>=4.0.0